server/requirements.txt
django==3.0.4
djangorestframework==3.11.0
graphene-django==2.9.0
graphene-django-extras==0.4.8
django-graphql-jwt==0.3.0
django-cors-headers==3.2.1